requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
assemblyai==0.37.0
google-generativeai==0.8.4
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.2.0
django-cors-headers==4.7.0
django-debug-toolbar

